+++ WindowsInspector.Controller/InspectorDriver.hpp
@@ -20,5 +20,5 @@ public:
 	InspectorDriver();
 	InspectorDriver(OwnedHandle hDriver);
 
-	SIZE_T ReadEvents(PVOID buffer, SIZE_T length);
+	DWORD ReadEvents(PVOID buffer, DWORD length);
 };
